added functionality to add point on click of the canvas

+++ src/main/java/org/miniworld/miniworld/Point.java
@@ -5,7 +5,7 @@ import java.util.Objects;
 public class Point {
     double x, y;
 
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
